+++ replication/src/change.rs
@@ -1,6 +1,6 @@
-use std::{marker::PhantomData, ops::{Deref, DerefMut}};
+use std::marker::PhantomData;
 use std::cell::UnsafeCell;
-use bevy::{ecs::{component::*, ptr::*, query::*, storage::*, system::SystemParam}, prelude::*};
+use bevy::{ecs::{component::*, ptr::*, query::*, storage::*}, prelude::*};
 use crate::*;
 
 /// Metadata about network-replicated types.
@@ -152,52 +152,4 @@ fn debug_unchecked_unwrap<T>(item: Option<T>) -> T {
     { item.unwrap() }
     #[cfg(not(debug_assertions))]
     { item.unwrap_unchecked() }
-}
-
-/// Immutable [`Resource`] access with replication metadata.
-#[derive(SystemParam)]
-pub struct NetRes<'w, T: ReplicableResource> {
-    /// Inner resource data.
-    pub data: Res<'w, T>,
-    /// Resource metadata.
-    pub meta: Res<'w, ReplicateMeta<T>>,
-}
-
-impl<'w, T: ReplicableResource> NetRes<'w, T> {
-
-}
-
-impl<T: ReplicableResource> Deref for NetRes<'_, T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.data
-    }
-}
-
-/// Mutable [`Resource`] access with replication metadata.
-#[derive(SystemParam)]
-pub struct NetResMut<'w, T: ReplicableResource> {
-    /// Inner resource data.
-    pub data: ResMut<'w, T>,
-    /// Resource metadata.
-    pub meta: Res<'w, ReplicateMeta<T>>,
-}
-
-impl<'w, T: ReplicableResource> NetResMut<'w, T> {
-
-}
-
-impl<T: ReplicableResource> Deref for NetResMut<'_, T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.data
-    }
-}
-
-impl<T: ReplicableResource> DerefMut for NetResMut<'_, T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.data
-    }
 }
